requirements.txt
torch
pandas
psutil
py-cpuinfo
tinydb
flask
request
flask_classful